succesfully transferred code over to activity and class pages
